Moved _set_time_axis() up to parent Decomposer class and added stub file for tests.
